Add files via upload
I added a diagram for an idea for what the front page will look like. Sections of the page that have functionality are color coded
Anything colored in Orange will just hold a picture 
Anything colored in Green will have a function

A Few Ideas
In the navigation bar on the upper right there are three buttons.
The home button will be a link back to the home page
The Cart button will be a link to the shopping cart page
The $ button will be a link to the crypto donation page (This will be an optional objective. If we get there we get there)

In the grid space with the green halves, there will be some kind of button to add items to the customer cart. (I will have to do research on this function)